requirements/requirements.txt
python-okx==0.2.3
python-binance==1.0.19
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.2
TA_Lib==0.4.28